+++ killer_instagram/src/database/db.py
@@ -1,19 +1,16 @@
+from contextlib import contextmanager
 
 from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy.orm import sessionmaker, Session
 
 from src.conf.config import settings
 
-"""
-Тут не потрібно міняти
-"""
 
 SQLALCHEMY_DATABASE_URL = settings.sqlalchemy_database_url
 print(SQLALCHEMY_DATABASE_URL)
 engine = create_engine(SQLALCHEMY_DATABASE_URL)
 
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
-
 
 # Dependency
 def get_db():
@@ -22,3 +19,14 @@ def get_db():
         yield db
     finally:
         db.close()
+
+@contextmanager
+def db_transaction(session: Session):
+    try:
+        yield session
+        session.commit()
+    except Exception as e:
+        session.rollback()
+        raise e
+    finally:
+        session.close()
